refactor: 메인 페이지 배경 아트워크 및 타이포 컴포넌트 분리
